Try to unbreak the tinderbox: Cast a time_t argument to long before printing it with %ld.
Former-commit-id: 5cbd95b3af9b40762afe403228dd7483e4beb33d

+++ sys/netatm/atm_sys.h
@@ -509,7 +509,7 @@ define|#
 directive|define
 name|ATM_TIME
 define|\
-value|struct timeval now, delta;					\ 	KT_TIME(now);							\ 	delta.tv_sec = now.tv_sec - atm_debugtime.tv_sec;		\ 	delta.tv_usec = now.tv_usec - atm_debugtime.tv_usec;		\ 	atm_debugtime = now;						\ 	if (delta.tv_usec< 0) {					\ 		delta.tv_sec--;						\ 		delta.tv_usec += 1000000;				\ 	}								\ 	printf("%3ld.%6ld: ", delta.tv_sec, delta.tv_usec);
+value|struct timeval now, delta;					\ 	KT_TIME(now);							\ 	delta.tv_sec = now.tv_sec - atm_debugtime.tv_sec;		\ 	delta.tv_usec = now.tv_usec - atm_debugtime.tv_usec;		\ 	atm_debugtime = now;						\ 	if (delta.tv_usec< 0) {					\ 		delta.tv_sec--;						\ 		delta.tv_usec += 1000000;				\ 	}								\ 	printf("%3ld.%6ld: ", (long)delta.tv_sec, delta.tv_usec);
 end_define
 
 begin_define
